API

BLR.binModel — Function

bin the model into a histogram, where each bin is the integrated value of the yVariable as a function of the xVariable params: m: model - model object to bin yVariable: Union{String, Symbol, Function} – if String then will be converted to symbols - y variable to bin - must be valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) bins: Union{Int, Vector{Float64}} - number of bins or bin edges for binning - if number of bins, then the bins will be equally spaced between the min and max of the xVariable - if bin edges then the number of bins will be length(bins)-1 - leftmost edge is then the first bin and the rightmost edge being the last bin - left edge of each bin is inclusive, right edge is exclusive except for the last bin which is inclusive xVariable: Union{String, Symbol, Function} - variable to bin over - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings dx: each bin is an integral over the range of xVariable – if dx is provided this is used as the associated integral element, otherwise defaults to ΔA in each ring struct returns: (binEdges, binCenters): Tuple{Vector{Float64},Vector{Float64}} - bin edges and centers for the xVariable of the histogram yBinned: Vector{Vector{Float64}} - binned values of the yVariables - length of yBinned is the same as the number of yVariables, with each element being a vector of length equal to the number of bins

BLR.binModel — Method

bin the model into a histogram, where each bin is the integrated value of the yVariable as a function of the xVariable params: m: model - model object to bin yVariable: Union{String, Symbol, Function} – if String then will be converted to symbols - y variable to bin - must be valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) bins: Union{Int, Vector{Float64}} - number of bins or bin edges for binning - if number of bins, then the bins will be equally spaced between the min and max of the xVariable - if bin edges then the number of bins will be length(bins)-1 - leftmost edge is then the first bin and the rightmost edge being the last bin - left edge of each bin is inclusive, right edge is exclusive except for the last bin which is inclusive xVariable: Union{String, Symbol, Function} - variable to bin over - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings dx: each bin is an integral over the range of xVariable – if dx is provided this is used as the associated integral element, otherwise defaults to ΔA in each ring struct returns: (binEdges, binCenters): Tuple{Vector{Float64},Vector{Float64}} - bin edges and centers for the xVariable of the histogram yBinned: Vector{Vector{Float64}} - binned values of the yVariables - length of yBinned is the same as the number of yVariables, with each element being a vector of length equal to the number of bins

BLR.getVariable — Method

retrieve elements from model object and stack them into matrices for easy manipulation params: m: model - model object to extract variables from variable: Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) returns: y: Matrix{Float64} - matrix of extracted variable from model.rings, created by stacking the output variable for each ring - for example, if variable given is :I y will have shape (length(r), length(ϕ)) as at each r and ϕ there is a value of I

BLR.getVariable — Method

retrieve elements from model object and stack them into matrices for easy manipulation params: m: model - model object to extract variables from variable: Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) returns: y: Matrix{Float64} - matrix of extracted variable from model.rings, created by stacking the output variable for each ring - for example, if variable given is :I y will have shape (length(r), length(ϕ)) as at each r and ϕ there is a value of I

BLR.getVariable — Method

retrieve elements from model object and stack them into matrices for easy manipulation params: m: model - model object to extract variables from variable: Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) returns: y: Matrix{Float64} - matrix of extracted variable from model.rings, created by stacking the output variable for each ring - for example, if variable given is :I y will have shape (length(r), length(ϕ)) as at each r and ϕ there is a value of I

BLR.image! — Method

BLR.image(m::model,variable::Union{String,Symbol,Function},kwargs...) Generate an image of the model where the color of each point is determined by the variable provided. params: m: model - model object to extract variable from variable (optional): Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to :I (intensity) kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot

BLR.image! — Method

BLR.image(m::model,variable::Union{String,Symbol,Function},kwargs...) Generate an image of the model where the color of each point is determined by the variable provided. params: m: model - model object to extract variable from variable (optional): Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to :I (intensity) kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot

BLR.image — Method

BLR.image(m::model,variable::Union{String,Symbol,Function},kwargs...) Generate an image of the model where the color of each point is determined by the variable provided. params: m: model - model object to extract variable from variable (optional): Union{String, Symbol, Function} - variable to extract from model (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to :I (intensity) kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot

BLR.plot3d! — Method

Plot3d(m::model, variable::Union{String,Symbol,Function}=nothing, annotatedCamera::Bool=true,kwargs...) Generate a 3D plot of the model geometry, optionally colored by a variable. params: m: model - model object to plot variable (optional): Union{String, Symbol, Function} - variable to color the points by (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to nothing (no coloring) annotatedCamera (optional): Bool – default true - if true, will annotate the camera position and orientation in the plot kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot - 3D plot of the model geometry, optionally colored by the variable provided

BLR.plot3d! — Method

Plot3d(m::model, variable::Union{String,Symbol,Function}=nothing, annotatedCamera::Bool=true,kwargs...) Generate a 3D plot of the model geometry, optionally colored by a variable. params: m: model - model object to plot variable (optional): Union{String, Symbol, Function} - variable to color the points by (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to nothing (no coloring) annotatedCamera (optional): Bool – default true - if true, will annotate the camera position and orientation in the plot kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot - 3D plot of the model geometry, optionally colored by the variable provided

BLR.plot3d — Method

Plot3d(m::model, variable::Union{String,Symbol,Function}=nothing, annotatedCamera::Bool=true,kwargs...) Generate a 3D plot of the model geometry, optionally colored by a variable. params: m: model - model object to plot variable (optional): Union{String, Symbol, Function} - variable to color the points by (if String will be converted to Symbol) - must be a valid attribute of model.rings (e.g. :I, :v, :r, :e, :i, :ϕ) or a function that can be applied to model.rings - example: Keplerian disk time delays could be calculated like t(ring) = ring.r*(1 .+ sin.(ring.ϕ).*ring.i)) - if not provided, defaults to nothing (no coloring) annotatedCamera (optional): Bool – default true - if true, will annotate the camera position and orientation in the plot kwargs...: keyword arguments for Plots.plot returns: p: Plots.plot - 3D plot of the model geometry, optionally colored by the variable provided

BLR.t — Method

time delays for a point in a disk params: ring: ring - ring object to calculate time delays for returns: t: Array{Float64,} - time delays for each ring in the model